Docker contianer + FFmpeg + vid.stab support
Reworking the entire Makefile to support installation of all necessary
binaries to occur within the Docker container itself so that the host
system would only be responsible for Docker only.

This means that the installation of `ffmpeg` and binaries
and including the necessary vid.stab (video stablization)
binaries and how they interact are taken care of within the
Docker container only.

The `Makefile` will now only be responsible for starting, entering, and
destroying all aspects of the Docker container itself.

The more important piece of this is that anything that is created within
the Docker container will be avaiable to the host system via the
`media_drive/` Docker volume which is a bi-directional volume mount.
This allows the stablized videos and results of the work to be available on
the host system without much effort.

All the scripts that are now responsible for:
- resizing all images
- creating the video of the images in sequence
- stablize the video and export it out

now exist within the `/scripts` directory and are executed one after
another via a master script located on the root directory via:
`convert_images_to_stablized_video.sh`.

Unfortunately as of this commit, I have not been able to figure out how
to execute `./setup.sh` within the Docker container, so the README
includes instructions on needing to execute this within the Docker
container before anything else can be executed.